PipePackage added

+++ TMK_plus_plus/Forms/MainForm.cs
@@ -12,6 +12,7 @@ namespace TMK_plus_plus
     {
         private DataTable dt = new DataTable();
         private BindingSource bs = new();
+        private BindingSource bsPackage = new();
         private readonly FilterDgv filterDgv = new FilterDgv();
 
         public MainForm()
@@ -31,6 +32,7 @@ namespace TMK_plus_plus
                 row["Lenght"] = pipe.Lenght;
                 row["Radius"] = pipe.Radius;
                 row["Weight"] = pipe.Weight;
+                row["Package"] = pipe.Package;
 
                 dt.Rows.Add(row);
             }
@@ -45,8 +47,8 @@ namespace TMK_plus_plus
             dt.Columns.Add("Lenght", typeof(int));
             dt.Columns.Add("Radius", typeof(int));
             dt.Columns.Add("Weight", typeof(int));
+            dt.Columns.Add("Package", typeof(PipePackage));
 
-            //bs.DataSource = typeof(Pipe);
             DataGridViewPipes.AutoGenerateColumns = false;
             DataGridViewPipes.DataSource = bs;
 
@@ -56,6 +58,13 @@ namespace TMK_plus_plus
             ColumnLenght.DataPropertyName = nameof(Pipe.Lenght);
             ColumnRadius.DataPropertyName = nameof(Pipe.Radius);
             ColumnWeight.DataPropertyName = nameof(Pipe.Weight);
+            ColumnPipePackage.DataPropertyName = nameof(Pipe.Package);
+
+            DataGridViewPackage.AutoGenerateColumns = false;
+            DataGridViewPackage.DataSource = bsPackage;
+
+            ColumnPackageNumber.DataPropertyName = nameof(PipePackage.Number);
+            ColumnDate.DataPropertyName = nameof(PipePackage.DateCreate);
 
             var bNumber = new Binding("Text", filterDgv, nameof(FilterDgv.Number));
             bNumber.DataSourceUpdateMode = DataSourceUpdateMode.OnPropertyChanged;
@@ -107,36 +116,47 @@ namespace TMK_plus_plus
 
         private async void ButtonChangePipe_Click(object sender, EventArgs e)
         {
-            using var editPipeForm = new EditPipeForm();
-            editPipeForm.Owner = this;
-            editPipeForm.Text = "Изменение трубы";
             Pipe p = new Pipe(dt.Rows[DataGridViewPipes.SelectedRows[0].Index].ItemArray!);
-            editPipeForm.CurrentPipe = p;
-            if (editPipeForm.ShowDialog() == DialogResult.OK)
+            if (p.Package == null)
             {
-                await Program.PipeReader.UpdatePipe(editPipeForm.CurrentPipe);
-                RefreshDataGrid();
+                using var editPipeForm = new EditPipeForm();
+                editPipeForm.Owner = this;
+                editPipeForm.Text = "Изменение трубы";
+                editPipeForm.CurrentPipe = p;
+                if (editPipeForm.ShowDialog() == DialogResult.OK)
+                {
+                    await Program.PipeReader.UpdatePipe(editPipeForm.CurrentPipe);
+                    RefreshDataGrid();
+                }
             }
-
+            else { MessageBox.Show("Нельзя удалить трубу, находящуюся в пакете!"); }
         }
 
         private async void ButtonDelete_Click(object sender, EventArgs e)
         {
             Pipe p = new Pipe(dt.Rows[DataGridViewPipes.SelectedRows[0].Index].ItemArray!);
-            await Program.PipeReader.DeletePipe(p);
-            RefreshDataGrid();
+            if (p.Package == null)
+            {
+                await Program.PipeReader.DeletePipe(p);
+                RefreshDataGrid();
+            }
+            else { MessageBox.Show("Нельзя удалить трубу, находящуюся в пакете!"); }
+
         }
 
         private void DataGridViewPipes_RowPrePaint(object sender, DataGridViewRowPrePaintEventArgs e)
         {
             if ((bool)DataGridViewPipes.Rows[e.RowIndex].Cells[1].Value == false)
             {
-                DataGridViewPipes.Rows[e.RowIndex].DefaultCellStyle.BackColor = Color.Red;
+                DataGridViewPipes.Rows[e.RowIndex].DefaultCellStyle.ForeColor = Color.Red;
+                DataGridViewPipes.Rows[e.RowIndex].DefaultCellStyle.SelectionForeColor = Color.LightPink;//Color.PaleVioletRed;
             }
         }
 
         private async void RefreshDataGrid()
         {
+            bsPackage.Clear();
+            var packages = await Program.PackageReader.GetPackages();
             var pipes = await Program.PipeReader.GetPipes();
             int countIsOk = 0;
             int totalWeight = 0;
@@ -147,7 +167,7 @@ namespace TMK_plus_plus
                 totalWeight += p.Weight;
             }
             bs.DataSource = dt;
-
+            packages.ForEach(p => bsPackage.Add(p));
             label1.Text = "Всего " + bs.Count + ", из них " + countIsOk +
                 " без брака. Общий вес: " + totalWeight + " кг.";
         }
@@ -155,16 +175,15 @@ namespace TMK_plus_plus
         private void _filterDgv_PropertyChanged(object? sender, PropertyChangedEventArgs e)
         {
             var filters = new List<string>();
-            bs.RemoveFilter();
             if (filterDgv.Number > 0)
                 filters.Add($"{nameof(Pipe.Number)} = {filterDgv.Number}");
 
             if (!String.IsNullOrEmpty(filterDgv.IsOk))
             {
                 if (filterDgv.IsOk == "Нет")
-                    filters.Add($"{nameof(Pipe.IsOk)} = True");
-                else
                     filters.Add($"{nameof(Pipe.IsOk)} = False");
+                else
+                    filters.Add($"{nameof(Pipe.IsOk)} = True");
             }
 
 
@@ -196,7 +215,86 @@ namespace TMK_plus_plus
 
         }
 
+        private async void ButtonAddPackage_Click(object sender, EventArgs e)
+        {
+            using var editPackageForm = new EditPackageForm();
+            editPackageForm.Owner = this;
+            editPackageForm.Text = "Добавление нового пакета";
+            editPackageForm.CurrentPackage = new PipePackage();
+            if (editPackageForm.ShowDialog() == DialogResult.OK)
+            {
+                await Program.PackageReader.AddPackage(editPackageForm.CurrentPackage);
+                RefreshDataGrid();
+            }
+        }
+
+        private async void ButtonChangePackage_Click(object sender, EventArgs e)
+        {
+            using var editPackageForm = new EditPackageForm();
+            editPackageForm.Owner = this;
+            editPackageForm.Text = "Изменение трубы";
+            PipePackage? p = ((PipePackage)DataGridViewPackage.SelectedRows[0].DataBoundItem).Clone() as PipePackage;
+            editPackageForm.CurrentPackage = p!;
+            if (editPackageForm.ShowDialog() == DialogResult.OK)
+            {
+                await Program.PackageReader.UpdatePackage(editPackageForm.CurrentPackage!);
+                RefreshDataGrid();
+            }
+        }
 
 
+        private async void ButtonDeletePackage_Click(object sender, EventArgs e)
+        {
+
+            await Program.PackageReader.DeletePackage((PipePackage)DataGridViewPackage.SelectedRows[0].DataBoundItem);
+            RefreshDataGrid();
+        }
+
+        private void DataGridViewPackage_SelectionChanged(object sender, EventArgs e)
+        {
+            if (DataGridViewPackage.SelectedRows.Count > 0)
+                foreach (DataGridViewRow row in DataGridViewPipes.Rows)
+                {
+                    object o = dt.Rows[row.Index].ItemArray[6];
+                    if (o != DBNull.Value && o != null)
+                        if (((PipePackage)o).Equals(DataGridViewPackage.SelectedRows[0].DataBoundItem))
+                        {
+                            row.DefaultCellStyle.BackColor = Color.Yellow;
+                            row.DefaultCellStyle.SelectionBackColor = Color.GreenYellow;
+                        }
+                        else
+                        {
+                            row.DefaultCellStyle.BackColor = Color.White;
+                            row.DefaultCellStyle.SelectionBackColor = Color.Empty;
+                        }
+                }
+        }
+
+        private async void ButtonAddToPackage_Click(object sender, EventArgs e)
+        {
+            Pipe p = new Pipe(dt.Rows[DataGridViewPipes.SelectedRows[0].Index].ItemArray!);
+            if (p.Package != null)
+                MessageBox.Show("Данная труба уже добавлена в пакет " + p.Package.Number);
+            else
+            {
+                PipePackage pp = (PipePackage)DataGridViewPackage.SelectedRows[0].DataBoundItem;
+                p.Package = pp;
+                await Program.PipeReader.UpdatePipe(p);
+                RefreshDataGrid();
+            }
+
+        }
+
+        private async void ButtonDeleteFromPackage_Click(object sender, EventArgs e)
+        {
+            Pipe p = new Pipe(dt.Rows[DataGridViewPipes.SelectedRows[0].Index].ItemArray!);
+            if (p.Package != null)
+            {
+                p.Package = null;
+                await Program.PipeReader.UpdatePipe(p);
+                RefreshDataGrid();
+            }
+            else { MessageBox.Show("Данная труба не добавлена в пакет!"); }
+        }
     }
 }
